+++ backend/routers/consumers.py
@@ -67,10 +67,6 @@ from fastapi import APIRouter, HTTPException, status
 from internal.auth.creation import CreateConsumerForm, create_consumer
 from internal.auth.middleware import ConsumerAuthDep
 from internal.database.dependency import database_dependency
-from internal.queries.admin_issue_reports import (
-    AsyncQuerier as AdminIssueReportsQuerier,
-)
-from internal.queries.admin_issue_reports import CreateAdminIssueReportParams
 from internal.queries.badge import AsyncQuerier as BadgeQuerier
 from internal.queries.badge import GetConsumerBadgesRow
 from internal.queries.consumer import AsyncQuerier as ConsumerQuerier
@@ -79,19 +75,9 @@ from internal.queries.consumer import (
     GetConsumersRow,
     UpdateConsumerParams,
 )
-from internal.queries.models import (
-    AdminIssueReport,
-    AdminIssueType,
-    Reservation,
-    SellerIssueReport,
-    SellerIssueType,
-)
+from internal.queries.models import Reservation
 from internal.queries.reservations import AsyncQuerier as ReservationsQuerier
 from internal.queries.reservations import GetConsumersReservationsFullRow
-from internal.queries.seller_issue_reports import (
-    AsyncQuerier as SellerIssueReportsQuerier,
-)
-from internal.queries.seller_issue_reports import CreateSellerIssueReportParams
 from pydantic import BaseModel
 
 router = APIRouter(prefix="/consumers", tags=["consumers"])
@@ -346,162 +332,3 @@ async def get_streaks(conn: database_dependency, consumer: ConsumerAuthDep) -> i
             break
 
     return streak_count
-
-
-class CreateSellerIssueReportForm(BaseModel):
-    """Seller issue report creation form."""
-
-    issue_type: SellerIssueType
-    description: str
-
-
-@router.post(
-    "/me/reservations/{reservation_id}/report",
-    status_code=status.HTTP_201_CREATED,
-    summary="Create seller issue report",
-    description="Creates a new seller issue report for a specific reservation.",
-    tags=["reports"],
-)
-async def create_seller_issue_report(
-    reservation_id: int,
-    form: CreateSellerIssueReportForm,
-    conn: database_dependency,
-    consumer: ConsumerAuthDep,
-) -> SellerIssueReport:
-    """Create seller issue report.
-
-    Args:
-        reservation_id: reservation id from path
-        form: seller issue report creation form
-        conn: database connection
-        consumer: consumer session
-
-    Returns:
-        created seller issue report
-
-    Raises:
-        HTTPException: if failed to create report or reservation not owned by consumer
-    """
-    reservation = await ReservationsQuerier(conn).get_reservation(
-        reservation_id=reservation_id
-    )
-    if not reservation or reservation.consumer_id != consumer.user_id:
-        raise HTTPException(
-            status_code=status.HTTP_404_NOT_FOUND,
-            detail="Reservation not found or not owned by consumer",
-        )
-
-    report = await SellerIssueReportsQuerier(conn).create_seller_issue_report(
-        CreateSellerIssueReportParams(
-            reservation_id=reservation_id,
-            issue_type=form.issue_type,
-            description=form.description,
-        )
-    )
-    if not report:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Failed to create seller issue report",
-        )
-    return report
-
-
-class CreateAdminIssueReportForm(BaseModel):
-    """Admin issue report creation form."""
-
-    issue_type: AdminIssueType
-    description: str
-
-
-@router.post(
-    "/me/reports/admin",
-    status_code=status.HTTP_201_CREATED,
-    summary="Create admin issue report",
-    description="Creates a new admin issue report.",
-    tags=["reports"],
-)
-async def create_admin_issue_report(
-    form: CreateAdminIssueReportForm,
-    conn: database_dependency,
-    consumer: ConsumerAuthDep,
-) -> AdminIssueReport:
-    """Create admin issue report.
-
-    Args:
-        form: admin issue report creation form
-        conn: database connection
-        consumer: consumer session
-
-    Returns:
-        created admin issue report
-
-    Raises:
-        HTTPException: if failed to create report
-    """
-    report = await AdminIssueReportsQuerier(conn).create_admin_issue_report(
-        CreateAdminIssueReportParams(
-            user_id=consumer.user_id,
-            issue_type=form.issue_type,
-            description=form.description,
-        )
-    )
-    if not report:
-        raise HTTPException(
-            status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
-            detail="Failed to create admin issue report",
-        )
-    return report
-
-
-@router.get(
-    "/me/reports/admin",
-    status_code=status.HTTP_200_OK,
-    summary="Get admin issue reports",
-    description="Retrieves a list of admin issue reports created by the consumer.",
-    tags=["reports"],
-)
-async def get_admin_issue_reports(
-    conn: database_dependency, consumer: ConsumerAuthDep
-) -> list[AdminIssueReport]:
-    """Get admin issue reports.
-
-    Args:
-        conn: database connection
-        consumer: consumer session
-
-    Returns:
-        list of admin issue reports
-    """
-    return [
-        r
-        async for r in AdminIssueReportsQuerier(conn).get_admin_issue_reports_by_user(
-            user_id=consumer.user_id
-        )
-    ]
-
-
-@router.get(
-    "/me/reports/seller",
-    status_code=status.HTTP_200_OK,
-    summary="Get seller issue reports",
-    description="Retrieves a list of seller issue reports created by the consumer.",
-    tags=["reports"],
-)
-async def get_seller_issue_reports(
-    conn: database_dependency, consumer: ConsumerAuthDep
-) -> list[SellerIssueReport]:
-    """Get seller issue reports.
-
-    Args:
-        conn: database connection
-        consumer: consumer session
-
-    Returns:
-        list of seller issue reports
-    """
-    return [
-        r
-        async for r in SellerIssueReportsQuerier(conn).get_seller_issue_reports_by_user(
-            consumer_id=consumer.user_id
-        )
-    ]
